Latest version as of 5am 8/2/19
index.html is the current version of the UI, It automatically detects the user's MetaMask account and displays different controls depending on whether the users is a student, admin, owner, or is not registered. In the student view, it automatically displays the student's balance, total traded, and reputation, and dynamically updates them every 3 seconds. It also automatically displays a list of transactions, and adds new transactions automatically.  I have also done some graphic design on it.

Also included are two earlier versions, one with all the same functionality, but without the styling, and one which also lacks the dynamic updating and separation of roles.